+++ mlxtend/classifier/__init__.py
@@ -4,17 +4,22 @@
 #
 # License: BSD 3 clause
 
-from .perceptron import Perceptron
 from .adaline import Adaline
-from .logistic_regression import LogisticRegression
-from .softmax_regression import SoftmaxRegression
-from .multilayerperceptron import MultiLayerPerceptron
 from .ensemble_vote import EnsembleVoteClassifier
+from .logistic_regression import LogisticRegression
+from .multilayerperceptron import MultiLayerPerceptron
+from .oner import OneRClassifier
+from .perceptron import Perceptron
+from .softmax_regression import SoftmaxRegression
 from .stacking_classification import StackingClassifier
 from .stacking_cv_classification import StackingCVClassifier
 
-__all__ = ["Perceptron", "Adaline",
-           "LogisticRegression", "SoftmaxRegression",
+__all__ = ["Adaline",
+           "EnsembleVoteClassifier",
+           "LogisticRegression",
            "MultiLayerPerceptron",
-           "EnsembleVoteClassifier", "StackingClassifier",
+           "OneRClassifier",
+           "Perceptron",
+           "SoftmaxRegression",
+           "StackingClassifier",
            "StackingCVClassifier"]